+++ OpenDental/Bridges/PaperlessTechnology.cs
@@ -67,7 +67,7 @@ namespace OpenDental.Bridges{
 			for(int i=0;i<fieldVals.Length;i++) {
 				fieldVals[i]=fieldVals[i].Replace("\"","");//remove quotes
 			}
-			int patNum=PIn.PInt(GetVal("PAT_OTHID"));
+			long patNum=PIn.PInt(GetVal("PAT_OTHID"));
 			if(patNum==0){
 				MessageBox.Show(patNum.ToString()+" is not a recognized PatNum.");
 				return;
@@ -315,7 +315,7 @@ namespace OpenDental.Bridges{
 					return;
 				}
 				fieldNames=lines[0].Split(',');
-				int diseaseDefNum;
+				long diseaseDefNum;
 				Disease disease;
 				string diseaseNote;
 				for(int i=1;i<lines.Length;i++){
